back/requirements.txt
asgiref
Django
sqlparse
django-extensions
langchain_aws
djangorestframework
djangorestframework-simplejwt
markdown
django-filter
drf-nested-routers
django-cors-headers
django-environ
django-allauth
cryptography

pytest
pytest-describe
mockito
pytest-django